+++ recurring_request.go
@@ -1,30 +1,5 @@
 package mcpayment
 
-// TokenizeRegisterReq body request for register token
-type TokenizeRegisterReq struct {
-	RegisterID  string    `json:"register_id" valid:"required,length(1|100)"`
-	CallbackURL string    `json:"callback_url" valid:"required,url"`
-	ReturnURL   string    `json:"return_url" valid:"required,url"`
-	IsTrx       bool      `json:"is_transaction,omitempty" valid:"required"`
-	TrxDetail   TrxDetail `json:"transaction,omitempty" valid:"-"`
-}
-
-// TrxDetail body request for transactiond detail
-type TrxDetail struct {
-	Amount float64 `json:"amount" valid:"-"`
-	Desc   string  `json:"description" valid:"-"`
-}
-
-// TokenizeGetReq param for get tokenize
-type TokenizeGetReq struct {
-	RegisterID string `json:"register_id" valid:"required"`
-}
-
-// TokenizeDelReq param for delete tokenize
-type TokenizeDelReq struct {
-	RegisterID string `json:"register_id" valid:"required"`
-}
-
 // RecrCreateReq body request for create recurring
 type RecrCreateReq struct {
 	RegisterID         string           `json:"register_id" valid:"required,length(1|100)"`
